refactor: update js interface
