--- src/NativeSourcepointCmp.ts
@@ -1,5 +1,0 @@
-import { TurboModuleRegistry } from 'react-native';
-
-import type { Spec } from './types';
-
-export default TurboModuleRegistry.getEnforcing<Spec>('RNSourcepointCmp');
